rating object primary key type is changed...

+++ app/src/main/java/com/example/openmarket/data/Rating.kt
@@ -10,5 +10,5 @@ class Rating(
     @ColumnInfo(name = "username") var username: String,
     @ColumnInfo(name = "rate_no") var rateNo: Double,
     @ColumnInfo(name = "product_id") var product_id: Long,
-    @PrimaryKey var id: Int = 0
+    @PrimaryKey var id: Long = 0
 ) : Serializable
